WilsonServer: Moved more files into 'common'
