+++ src/wilson/types/ServerEvent.ts
@@ -1,8 +1,8 @@
 import EServerAction from "../enums/EServerAction";
-import ServerNewConversationInviteEventPayload from "./events/payloads/server/ServerNewConversationInviteEventPayload";
-import ServerNewMessageEventPayload from "./events/payloads/server/ServerNewMessageEventPayload";
-import ServerWelcomeEventPayload from "./events/payloads/server/ServerWelcomeEventPayload";
-import IServerEvent from "./events/ServerEvent";
+import ServerNewConversationInviteEventPayload from "./common/events/payloads/server/ServerNewConversationInviteEventPayload";
+import ServerNewMessageEventPayload from "./common/events/payloads/server/ServerNewMessageEventPayload";
+import ServerWelcomeEventPayload from "./common/events/payloads/server/ServerWelcomeEventPayload";
+import IServerEvent from "./common/events/ServerEvent";
 
 export type ServerWelcomeEvent = IServerEvent<ServerWelcomeEventPayload, EServerAction.WELCOME>;
 export type ServerNewMessageEvent = IServerEvent<ServerNewMessageEventPayload, EServerAction.NEW_MESSAGE>;
