--- src/wilson/types/events/ClientEvent.ts
@@ -1,8 +1,0 @@
-import EClientAction from "../../enums/EClientAction";
-
-interface IClientEvent<T, U extends EClientAction> {
-    action: U,
-    payload: T
-}
-
-export default IClientEvent;
