--- src/wilson/types/events/ServerEvent.ts
@@ -1,9 +1,0 @@
-import EServerAction from "../../enums/EServerAction";
-
-interface IServerEvent<T, U extends EServerAction> {
-    to: string,
-    action: U,
-    payload: T
-}
-
-export default IServerEvent;
