--- src/wilson/types/events/payloads/client/ClientNewConversationInviteAckEventPayload.ts
@@ -1,8 +1,0 @@
-import ConversationInviteID from "../../../../types/common/ConversationInviteID";
-
-interface ClientNewConversationInviteAckEventPayload {
-    invite_id: ConversationInviteID
-    accept: boolean;
-}
-
-export default ClientNewConversationInviteAckEventPayload;
