--- src/wilson/types/events/payloads/client/ClientNewConversationInviteEventPayload.ts
@@ -1,7 +1,0 @@
-import UserID from "../../../../types/common/UserID";
-
-interface ClientNewConversationInviteEventPayload {
-    participants: UserID[]
-}
-
-export default ClientNewConversationInviteEventPayload;
