--- src/wilson/types/events/payloads/client/ClientNewMessageEventPayload.ts
@@ -1,8 +1,0 @@
-import UserID from "../../../../types/common/UserID";
-
-interface ClientNewMessageEventPayload {
-    to: UserID;
-    message: string
-}
-
-export default ClientNewMessageEventPayload;
