--- src/wilson/types/events/payloads/server/ServerNewConversationEventPayload.ts
@@ -1,6 +1,0 @@
-interface ServerNewConversationEventPayload {
-    conversation_id: string,
-    participants: string[]
-}
-
-export default ServerNewConversationEventPayload;
