--- src/wilson/types/events/payloads/server/ServerNewConversationInviteEventPayload.ts
@@ -1,7 +1,0 @@
-import ConversationInvite from "../../../../types/common/ConversationInvite";
-
-interface ServerNewConversationInviteEventPayload {
-    invite: ConversationInvite
-}
-
-export default ServerNewConversationInviteEventPayload;
