--- src/wilson/types/events/payloads/server/ServerNewMessageEventPayload.ts
@@ -1,6 +1,0 @@
-interface ServerNewMessageEventPayload {
-    from: string,
-    message: string
-}
-
-export default ServerNewMessageEventPayload;
