--- src/wilson/types/events/payloads/server/ServerWelcomeEventPayload.ts
@@ -1,6 +1,0 @@
-interface ServerWelcomeEventPayload {
-    client_id: string;
-    server_name: string;
-}
-
-export default ServerWelcomeEventPayload;
